PyDownload-master/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
idna==2.9
pytube3==9.6.4
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.9
